+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the chef_auto_accumulator cookbook.
 
+## Unreleased
+
+- Add an option to sort Array configuration values
+
 ## 0.4.3 - *2024-05-06*
 
 ## 0.4.2 - *2024-05-06*